MSound/setSeExtParameter: 39.2% -> 41.5% (restore SE volume scaling)

+++ include/MSound/MSRandVol.hpp
@@ -10,7 +10,7 @@ class MSRandVol {
 public:
 	MSRandVol(u32);
 
-	virtual u32 getRandVol(u32);
+	virtual f32 getRandVol(u32);
 
 	static void construct(u32);
 	u32 getRandomVolume(u32, u32);
